Release: Version fourier-cpp 0.6.0, fourier 0.6.0; fix: velocity and torque feedforward

+++ fourier/include/aios.h
@@ -206,8 +206,8 @@ typedef struct Feedback_ *FourierFeedbackPtr;
 
 typedef struct PosPtInfo_ {
   float pos;
-  short vel;
-  short torque;
+  float vel;
+  float torque;
 } PosPtInfo;
 
 typedef struct Command_ {
